requirements-gpu.txt
black
cdifflib
datasets
flake8
huggingface_hub[cli]
huggingface_hub[hf_xet]
ipykernel
ipython
ipywidgets
isort
jupyterlab
librosa
loguru
matplotlib
notebook
nltk
numpy<2
pydantic>=2
pandas
peft>= 0.6.0
pillow
pip
pytest
python-dotenv
scikit-learn
seaborn
sentencepiece>=0.1.99
streamlit
sortedcontainers
soundfile
textstat
torch==2.6.0
torchvision
transformers
trl
tqdm
typer
wandb
wordcloud
-e .